// Ref/Web/ReferenceDataUpdateService.Web/src/models/RefAccTaxRateUserView.ts
//------------------------------------------------------------------------------
// <auto-generated>
//     This code was generated from a template.
//
//     Manual changes to this file may cause unexpected behavior in your application.
//     Manual changes to this file will be overwritten if the code is regenerated.
// </auto-generated>
//------------------------------------------------------------------------------
import { IEntity } from "./IEntity";

export class RefAccTaxRateUserView implements IEntity {
	ZAT_PK: string;
	ZAT_RN_NKCountry: string;
	static readonly ZAT_RN_NKCountry_MaxLength: number = 2;
	ZAT_ReferenceRateType: string;
	static readonly ZAT_ReferenceRateType_MaxLength: number = 10;
	ZAT_StartDate: string;
	ZAT_EndDate: string;
	ZAT_RateNumerator: number;
	ZAT_RateDenominator: number;
	ZAT_IsSystem: boolean;
	ZAT_IsPublished: boolean;
	ZAT_IsEditable: boolean;
}